Separate CLI runtime responsibilities
